Add a BackNavigationScreen composable and use as a base for SettingsScreen and add settings directly to the screen composable
